Conflict between `nup' and `addtotoc' fixed.
This bug was introduces in version 0.4x. A \newpage was inserted after
each page, assuming that it does not trigger a new page as long as
nothing was typeset onto this page. This assumtion was wrong. In macro
\AM@addtotoc some content seems to be typeset (maybe a space?) by \@sect.

Bug was fixed by introducing \ifAM@newpage and inserting \newpage only
when a new page is really needed.